refactor(agents): relocate credentials/reconcile/seed; auth_config becomes auth
