+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -6,12 +6,12 @@ use std::sync::Arc;
 use crate::adapters::git::WorkspaceFileSearchCache;
 use crate::adapters::processes::ProcessService;
 use crate::api::auth::AuthManager;
-use crate::domains::agents::auth_config::{AgentAuthConfigService, AgentAuthConfigStore};
+use crate::domains::agents::auth::{AgentAuthConfigService, AgentAuthConfigStore};
 use crate::domains::agents::model_registry::service::DynamicModelRegistryService;
 use crate::domains::agents::model_registry::store::DynamicModelRegistryStore;
-use crate::domains::agents::reconcile::execution::AgentReconcileService;
+use crate::domains::agents::installer::reconcile::execution::AgentReconcileService;
 use crate::domains::agents::runtime::AgentRuntime;
-use crate::domains::agents::seed::AgentSeedStore;
+use crate::domains::agents::installer::seed::AgentSeedStore;
 use crate::domains::artifacts::protection::ArtifactProtectionService;
 use crate::domains::artifacts::runtime::ArtifactRuntime;
 use crate::domains::cowork::artifacts::CoworkArtifactRuntime;
