+++ anyharness/crates/anyharness-lib/src/domains/agents/auth/claude_tests.rs
