+++ anyharness/crates/anyharness-lib/src/domains/agents/auth/scope_tests.rs
