+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/agent_process.rs
@@ -8,7 +8,7 @@ use crate::domains::agents::readiness::paths::{
     artifact_root, has_managed_registry_binary_for_names,
 };
 use crate::domains::agents::registry::built_in_registry;
-use crate::domains::agents::seed;
+use crate::domains::agents::installer::seed;
 use crate::integrations::agent_cli::acp_registry::{self, ResolvedRegistryDistribution};
 use crate::integrations::agent_cli::executable::{find_real_binary_in_path, is_valid_executable};
 use crate::integrations::agent_cli::launcher::generate_launcher_script;
