+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/reconcile/execution.rs
@@ -8,7 +8,7 @@ use uuid::Uuid;
 use super::{reconcile_agent, AgentReconcileOutcome, AgentReconcileResult};
 use crate::domains::agents::installer::InstallOptions;
 use crate::domains::agents::model::{AgentDescriptor, AgentKind};
-use crate::domains::agents::seed::AgentSeedStore;
+use crate::domains::agents::installer::seed::AgentSeedStore;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum AgentReconcileJobStatus {
