+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/seed/archive.rs
