+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/service.rs
@@ -4,7 +4,7 @@ use super::agent_process::{self, install_agent_process_artifact, is_agent_proces
 use super::lock::AgentInstallLock;
 use super::native::{install_native_artifact, is_native_installable};
 use crate::domains::agents::model::*;
-use crate::domains::agents::seed;
+use crate::domains::agents::installer::seed;
 use crate::integrations::agent_cli::launcher::LauncherError;
 
 #[derive(Debug, Clone)]
